Use fixed versions of NEON and Flatbuffers
Up until now the master branch was used from both repositories. This meant that
we were at risk of build errors due to upstream changes outside of our control.

There are no tags in the ARM_NEON_2_x86_SSE repository so we are selecting the
current latest commit as we know it works.

For Flatbuffers we are using the latest tag known to work, v1.9.0.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>