onnx-runtime: Add Google RE2 regular expression library.
This is used in new version of gRPC.

Signed-off-by: Clément Péron <peron.clem@gmail.com>

The onnx-runtime libraries use the RE2 library. However, it is only built as a dependency of FeaturizersLibrary,
so if you do not build FeaturizersLibrary by setting -Donnxruntime_USE_FEATURIZERS=OFF or using the default setting,
the build will fail.

This recipe has already been added to upstream meta-openembedded:
https://github.com/openembedded/meta-openembedded/blob/master/meta-oe/recipes-support/re2/re2_2020.11.01.bb

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>