onnx-runtime: Use re2 instead of relying on Featurizer option to build it
The onnx-runtime libraries use the RE2 library. However, it is only built as a dependency of FeaturizersLibrary,
so if you do not build FeaturizersLibrary by setting -Donnxruntime_USE_FEATURIZERS=OFF or using the default setting,
the build will fail.

Therefore, instead directly use the re2 library.

This allows us to remove a patch to work around a unit test build failure.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>