scripts: Apply BSP workarounds
In order to make a clean build using the RZ/G2L BSP there are two
required workarounds that should be applied. The first is that U-Boot must
be built before the trusted-firmware to avoid potential dependency issues.
The second is to apply a patch from poky to resolve -march/-mcpu conflicts
during compilation.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>