meta-benchmark: delegate: Add test model list for low memory devices
The RZ/G2LC has 1GB of DDR, 384MB of which is dedicated to CMA in the
DTS. As swapspace can only be used for pages which can be transferred
to that area, out of memory errors can occur in the delegate benchmark
when using the GPU as a backend. Providing a model list that is tested
for similar boards ensures the best possible coverage for this scenario.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>