gitlab-ci.yml: Add testing support
LAVA test jobs need to know the kernel/dtb/filesystem binaries to
download. These files are currently stored in GitLab CI artifacts from the
previous build jobs. In order for the subsequent test jobs to get the URLs
to the binaries the test jobs need to know the corresponding build job ID.

We do this by passing the $CI_JOB_ID variable into the job.env file which
gets saved as a special report in GitLab. This environment setting is then
passed on in the environment of the subsequent test job.

More details on this process can be found in GitLab's documentation:
https://docs.gitlab.com/ee/ci/environments/#set-dynamic-environment-urls-after-a-job-finishes

At the moment the test jobs only submit a LAVA job. They do not wait to
see if the test passes.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>