scripts: submit-lava-job.sh: Add ability to check LAVA test results
The script can now monitor and report back on test results from the
submitted LAVA test job.

The GitLab CI job will fail if the overall test job fails to complete.

Also changing the GitLab CI runner tag to "lava-job", which for us is a
special runner that can handle multiple concurrent jobs, given the CPU
load is so low for them.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>