gitlab-ci.yml: Temporarily make all ek874 tests manual
The RZ/G2E ek874 boards we currently have in the LAVA lab are RevA boards,
which are not currently supported in the latest RZ/G2 BSP and don't fully
boot.

Until we get newer boards, let's disable the automatic testing for ek874
to avoid continuous CI pipeline failures we cannot avoid.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>