scripts: submit-lava-job.sh: Fix options parsing
Fix a bug where the -r option fails to work if it isn't defined at the end
of the script argument list.

The issue is caused because there is no additional argument for the -r
option, but we were still shifting the argument list.

The same change has also been done for the -h argument.

Fixes: 5de3e44d4ac3 ("scripts: submit-lava-job.sh: Add ability to check LAVA test results")

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>