Upgrade to use RZ/G BSP-3.0.3
The latest version of the RZ/G BSP has changed how it uses bitbake
configuration files. In this commit we switch to using the new approach.

The build script is configured to use the following proprietary packages:
RZ/G2H family:
- RTK0EF0045Z0022AZJ-v1.0.1_EN.zip
RZ/G2L family:
- RTK0EF0045Z13001ZJ-v1.0.5_EN.zip
- RTK0EF0045Z15001ZJ-v1.1.0_EN.zip

If different versions are required, scripts/build-rzg-ai-bsp.sh will need
to be updated accordingly.

This commit also updates the documentation and CI configuration as
required to use the v3.0.3 BSP.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>